Refactoring PartitionFile.java further. Implemented PartitionFile interface and renamed original PartitionFile.java to PartitionFileImpl.java.
