+++ src/main/java/com/teragrep/cfe_39/consumers/kafka/PartitionFile.java
@@ -45,126 +45,14 @@
  */
 package com.teragrep.cfe_39.consumers.kafka;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import com.teragrep.cfe_39.Config;
-import com.teragrep.cfe_39.avro.SyslogRecord;
-import com.teragrep.cfe_39.consumers.kafka.queue.WritableQueue;
-import com.teragrep.rlo_06.ParseException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.ListIterator;
 
-public class PartitionFile {
+public interface PartitionFile {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(PartitionFile.class);
+    void addRecord(KafkaRecordImpl kafkaRecord);
 
-    private final String topic;
-    private final String partition;
-    private final Config config;
-    private final File syslogFile;
-    private final List<KafkaRecordImpl> kafkaRecordList;
-    private final List<Long> batchOffsets;
+    void commitRecords() throws IOException;
 
-    PartitionFile(Config config, String topic, String partition) throws IOException {
-        WritableQueue writableQueue = new WritableQueue(config.getQueueDirectory(), topic + partition);
-        this.syslogFile = writableQueue.getNextWritableFile();
-        this.kafkaRecordList = new ArrayList<>();
-        this.config = config;
-        this.topic = topic;
-        this.partition = partition;
-        this.batchOffsets = new ArrayList<>();
-    }
-
-    public void addRecord(KafkaRecordImpl kafkaRecord) {
-        kafkaRecordList.add(kafkaRecord);
-    }
-
-    public void commitRecords() throws IOException {
-        ListIterator<KafkaRecordImpl> kafkaRecordListIterator = kafkaRecordList.listIterator();
-        long storedOffset = 0;
-        while (kafkaRecordListIterator.hasNext()) {
-            KafkaRecordImpl next = kafkaRecordListIterator.next();
-            SyslogRecord syslogRecord;
-            try {
-                syslogRecord = next.toSyslogRecord();
-            }
-            catch (ParseException e) {
-                if (config.getSkipNonRFC5424Records()) {
-                    LOGGER
-                            .warn(
-                                    "Skipping parsing a non RFC5424 record, record metadata: <{}>. Exception information: ",
-                                    next.offsetToJSON(), e
-                            );
-                    JsonObject recordOffset = JsonParser.parseString(next.offsetToJSON()).getAsJsonObject();
-                    if (recordOffset.get("offset").getAsLong() > storedOffset) {
-                        storedOffset = recordOffset.get("offset").getAsLong();
-                    }
-                    continue;
-                }
-                else {
-                    LOGGER.error("Failed to parse RFC5424 record <{}>", next.offsetToJSON());
-                    throw new RuntimeException(e);
-                }
-            }
-            catch (NullPointerException e) {
-                if (config.getSkipEmptyRFC5424Records()) {
-                    LOGGER
-                            .warn(
-                                    "Skipping parsing an empty RFC5424 record, record metadata: <{}>. Exception information: ",
-                                    next.offsetToJSON(), e
-                            );
-                    JsonObject recordOffset = JsonParser.parseString(next.offsetToJSON()).getAsJsonObject();
-                    if (recordOffset.get("offset").getAsLong() > storedOffset) {
-                        storedOffset = recordOffset.get("offset").getAsLong();
-                    }
-                    continue;
-                }
-                else {
-                    LOGGER.error("Failed to parse RFC5424 record <{}> because of null content", next.offsetToJSON());
-                    throw new RuntimeException(e);
-                }
-            }
-            // When the file size has gone above the maximum, commit the file into HDFS using the latest topic/partition/offset values as the filename and then delete the local avro-file.
-            if (config.getMaximumFileSize() < syslogFile.length()) {
-                writeToHdfs(storedOffset);
-            }
-            // SyslogAvroWriter initialization will re-initialize the syslogFile if it has been deleted because of writeToHdfs().
-            try (SyslogAvroWriter syslogAvroWriter = new SyslogAvroWriter(syslogFile)) {
-                syslogAvroWriter.write(syslogRecord);
-            }
-            if (syslogRecord.getOffset() > storedOffset) {
-                storedOffset = syslogRecord.getOffset();
-            }
-        }
-        // Clear the kafkaRecordList from successfully committed records.
-        kafkaRecordList.clear();
-        // Store the last offset of the batch to a list.
-        if (storedOffset > 0) {
-            batchOffsets.add(storedOffset);
-        }
-    }
-
-    public void writeToHdfsEarly() throws IOException {
-        if (!batchOffsets.isEmpty()) {
-            writeToHdfs(batchOffsets.get(batchOffsets.size()-1));
-        }
-    }
-
-    // Writes the file to hdfs and initializes new file.
-    private void writeToHdfs(long offset) throws IOException {
-        try (HDFSWrite writer = new HDFSWrite(config, topic, partition, offset)) {
-            writer.commit(syslogFile); // commits the final AVRO-file to HDFS.
-        }
-        syslogFile.delete(); // Delete the file as all the contents have been stored to HDFS.
-        try (SyslogAvroWriter syslogAvroWriter = new SyslogAvroWriter(syslogFile)) {
-            // NoOp, syslogAvroWriter has initialized the empty AVRO-file.
-        }
-    }
+    void writeToHdfsEarly() throws IOException;
 
 }
